requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.8
openai>=1.0.0
pydantic>=1.10.0
PyPDF2>=3.0.0
pillow>=9.0.0
pytesseract>=0.3.10
pymupdf>=1.22.0 
fastapi
uvicorn
python-multipart